+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 12.8
 - Fix PHP warning undefined array key when switching Meals in ServeMenus.php
 - Fix SQL error when no Quarters found (Course widget) in Course.php
 - Fix PHP8.5 deprecated trim() Passing null to parameter 1 of type string in StudentFieldBreakdown.php & StudentBreakdown.php
+- Display Enrollment Calendar to Student in Calendar.php
 
 Changes in 12.7.3
 -----------------